Update the nix build configuration.
Remove the old shell.nix with some legacy versions pinned and replace it
with a flake-based shell. It installs rust via rustup instead of fenix
to be more generally compatible with the guidelines.

This also adds the rust-toolchain.toml spec with all the components
required for wasm, and everything else to make rust-analyzer & clippy
happy.